requirements.txt
discord.py>=2.3.0
wavelink>=2.6.0
python-dotenv>=1.0.0
spotipy>=2.23.0
yt-dlp>=2023.10.0

